agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
langchain-ollama==0.2.2
langchain-core==0.3.28
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
